Made the modular version with complete features as of now Now Mobile V3
